backend/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.22
python-dotenv==0.19.1
eventlet==0.33.0
werkzeug==2.0.2
dnspython==2.1.0